first version of branching for fast and slow image implementations
